Separate textBox from textStyle panel.
