+++ assets/src/edit-story/elements/text/index.js
@@ -71,6 +71,7 @@ export const panels = [
   PanelTypes.SIZE_POSITION,
   PanelTypes.LAYER_STYLE,
   PanelTypes.TEXT_STYLE,
+  PanelTypes.TEXT_BOX,
   PanelTypes.BORDER_RADIUS,
   PanelTypes.BORDER,
   PanelTypes.LINK,
